pyproject.toml
[project]
name = "01-fs-img-generation-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "python-dotenv>=1.0.0",
]
